# packages/nestjs-sdk/src/gen/types/HealthStatusDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type HealthStatusDto = {
    /**
     * @type boolean
    */
    ok: boolean;
    /**
     * @type string, date-time
    */
    timestamp: string;
};