# packages/nestjs-sdk/src/gen/types/ListQueryDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ListQueryDto = {
    /**
     * @minLength 1
     * @default 1
     * @type number | undefined
    */
    page?: number;
    /**
     * @minLength 1
     * @maxLength 100
     * @default 10
     * @type number | undefined
    */
    limit?: number;
    /**
     * @type string | undefined
    */
    search?: string;
};